requirements.txt
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
streamlit
requests==2.28.2
neo4j~=5.21.0
networkx~=3.2.1
langchain
aiohttp
langchain-openai
langchain-core
langchain-community
langsmith
pandas~=2.2.2
PyDriller~=2.6
PyGithub
py2neo
gunicorn==20.1.0
redis~=5.0.8
GitPython~=3.1.43